% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmGammaNet.R
\name{fit.glmGammaNet}
\alias{fit.glmGammaNet}
\title{Fit glmnet model for Gamma distributed response data.}
\usage{
fit.glmGammaNet(A, b, exponential.dist = FALSE, alpha.EN = 0.5,
  num_lambda = 100L, glm_type = 1L, max_iter = 100L,
  abs_tol = 1e-04, rel_tol = 0.01, normalize_grad = FALSE,
  k_fold = 5L, has_intercept = TRUE, k_fold_iter = 5L,
  min.lambda.ratio = 1e-04, ...)
}
\arguments{
\item{A}{The matrix of independent variables.}

\item{b}{The vector of response variables.}

\item{exponential.dist}{Parameter to determine whether we use the Exponential distribution (TRUE) or the Gamma distribution (FALSE).}

\item{alpha.EN}{The coefficient of elastic net regularizer (1 means lasso).}

\item{num_lambda}{Size of the lambda grid.}

\item{glm_type}{Type of glm model, 1 is exponential, 2 is gamma (not implemented yet).}

\item{max_iter}{Max number of iteration for the prox grad descent optimizer.}

\item{abs_tol}{Absolute error threshold for the pgd optimizer.}

\item{rel_tol}{Relative error threshold for the pgd optimizer (not used for vanilla PGD).}

\item{normalize_grad}{Swtich for whether to normalize the gradient or not.}

\item{k_fold}{The number of folds for cross validation.}

\item{has_intercept}{Parameter to determine if there is an intercept (TRUE) or not (FALSE).}

\item{k_fold_iter}{The number of iterations for the cross-validation.}

\item{min.lambda.ratio}{Minimum lambda ratio for cross-validation.}

\item{...}{Additional parameters.}
}
\value{
vector of optimal coefficient for the glm model
}
\description{
Fit glmnet model for Gamma distributed response data.
}
